Move AuthCredentialsDTO to domain folder
