+++ library-ifal/src/core/domain/dto/AuthCredentialsDTO.ts
